src/sensorsiesta-server/requirements.txt
Django==1.8.2
djangorestframework==3.4.6
rpyc==3.3.0